Compiles on windows
It crashed on windows so I had to debug it, and to be able to debug
it I had make it work with microsofts compiler, this is using the
free Microsoft Visual C++ 2010 Express.

The main fix in this commit is a missing CL_CALLBACK.

I couldn't quite get the microsoft compiler to work, the returned
information from 'cl_basic' is wrong so it's disabled for now.

To make it compile with cl.exe I had to move unused parameters to
after the variable declartions.
On windows using long is not such great idea since it will be
32 bits even on 64 bits compilations when that day arrives.
Here size_t and unsigned long is mixed that will work as long as
long and size_t is the same size, I don't know if that is true for
windows64. (mingw-gcc 4.4.0 warns a lot when compiling)