Supply senders env in enif_send if done by a scheduler thread
Env must be present when invoking enif_send from a scheduler thread.

Some workarounds for things that have changed in the nif implemenation.
It might be a good idea to keep the 'extra' functionality
in the helper library 'clu', and keep 'cl' as close to the spec as
possible.

Minor Mac Makefile fixes

I messed up when converting to windows cl compiler
And cc_subdiv now compiles on mac, the display isn't that great on these
drivers I have on my mac-mini.  Everything is blended for some reason